cuda: Use MIG device handle to get stats
If we are using MIG then we don't have permission to access the
overall device stats. We also don't care about the overall free
memory.

This uses the UUID passed to the container, which is the MIG device
when MIG is enabled and the GPU when not, to get the correct type of
handle.